fix: 修 3 处 Phase 1 验收发现的缺陷（1 编译错误 + 2 测试缺陷）
- resp: 数组解析循环变量 int 与 int64 不匹配导致编译失败
- store: TTL 剩余秒数改向上取整，对齐 Redis 语义（100ms TTL 应返回 1 而非 0）
- store_test: Expire 缺失 key 断言写反（行为正确时反而误报失败），修正断言方向
- resp_test: 测试输入 bulk 长度笔误 $4 改 $3（key 为 3 字节）

背景：T-101 验收（老板授权旧代码烂可以修改）；修后 go build/vet/test 全绿。

+++ internal/resp/resp.go
@@ -113,7 +113,7 @@ func (rd *Reader) readValue(t Type) (Value, error) {
 			return Value{}, errors.New("resp: invalid array length")
 		}
 		arr := make([]Value, n)
-		for i := 0; i < n; i++ {
+		for i := int64(0); i < n; i++ {
 			b, err := rd.r.ReadByte()
 			if err != nil {
 				return Value{}, err
